Fixes problem where get_or_create could return two objects.
For example, this could happen:
1.- Create application with applicant:
Dr John Smith Male
Career stage: Undergraduate / master's student
http://localhost:8000/proposal/ea2d7246-4221-467e-8110-d0531c672c14/update

2.- Create another application with applicant:
Dr John Smith Male
Career stage: PhD Student

There should be two John Smith in the PersonPosition database.

3.- Create application with overaraching project supervisor:
Dr John Smith

Because the get_or_create() for the project supervisor didn't have any
career stage: it was returning two objects. Now with the
career_stage=None it will create a new PersonPosition.

Another possible way would have been to choose the newest PersonPosition
or the highest ranking (what is Other anyway?). But we chose to be
strict and create a new one with the given information.

When we have Orcid we are going to solve problems with the
PhysicalPerson (unique identifier for this). For the PersonPosition even
with Orcid is unlikely to be solved any problem.