Fixes sending emails to the admin.
Previously I had used a @gmail.com for testing. Without the SERVER_EMAIL
Django uses root@localhost. When using @gmail.com it seems to replace it
by the account's email. But when using other services (my local postfix,
I think that the Microsoft Outlook online) it doesn't replace it and
emails are not delivered.