Fixes problem using the formset for the funding organisations and proposal partners.
Uses older jQuery for the formsets.

The problem was:
a) Add a funding organisation or a proposal partner and then press "add
another"
b) See that the new created funding organisation (for example) had a
non-working auto complete, an error on the Javascript console

More information:
https://github.com/elo80ka/django-dynamic-formset/issues/162

There are still UI problems: when clicking "add another" the original
django-autocomplete-light are converted to multiple selection lists.