Deletes call to jQuery.noConflict
This was added as part of the debugging (hopefully not as part of the
fixing) of the date time widget pickers for the proposal main form.

Having this made the Eligibility form Javascript (for removing/adding
the "*" for required text when Eligibility is "No") not work when the
Evaluation form was added in the Proposal Detail view (see previous
commits).