Makes "Eligibility" and "Edit/Create Evaluation" right hand side column appear only in the context of "Evaluation"
Previously all the ProposalDetail views showed it.

It also stops using Proposal UUIDs and starts using Proposal ID. The
UUIDs are used on the public side (outside management) to avoid people
guessing IDs. Internally is easier to have all with IDs for consistency
and in order to have the Comments/Attachments forms easier.