Fixes column spacing on the table of calls (specially when changing to new tabs)
Initially DataTables didn't know the size of its container for "Future"
and "Closed" because it was hidden (only the "Open" default tab). This
caused strange column sizes when moving to "Closed" or "Future". Now via
JS it recalculates the size of the columns so it makes sense.

Also, it has a fixed size of columns. Otherwise because the contents
changes (e.g. call names) could be potentially different for different
tabs.

fixes #281