Uses always "pk" instead of "id" on the URLs.
Earlier I moved many views from TemplateView to DetailView. DetailView
needs "pk". In the urls.py it was not consistent the use of id or pk.
Now it's always pk so there is less confusion.

All unit tests are passing but there might be some more instances of
this to be fixed.