Improves PartnerForm greately
Sops using self.fields['id'] as a way to pass the partner_id and uses
self.partner_id. This caused problems on the test code

Note that self.errors in a form is a property: it executes
self.full_clean(). This made len(self.errors) or using the debugger to
temporary fix the problem. Or calling self.full_clean also fixed the
problem.

For tomorrow: check that the partner is saved in the database