Invoices: fixes what happens when a new invoice is added
If the previous invoice had a file attached: the new invoice had (UI
only!) the file attached: checkbox to clear the file, change the file,
etc.

This fixes the UI. The UI is done via crispy-form. On crispy-form update
this should be double checked: if they change the layout or names it
might not work