Adds some validation in the Invoices or Financial Report
Note that it allows for now to create empty Financial Reports. This is
because it's using the same form for both and by default it shows one
entry. I'll re-evaluate this (e.g. ignoring empty forms as valid but not
creating anything in the database or splitting Invoices and Financial
Reports into two forms)