Fixes HTTP 500 if installment didnt' have amount
The 'amount' in an installment is required. But as per
https://docs.djangoproject.com/en/3.0/ref/forms/api/#django.forms.Form.use_required_attribute
the required HTML attribute is not added for forms in formsets so the
clean() methods of forms could receive incomplete data. This avoids
trying to validate forms that have errors. When the error (e.g.
"Required field") happens it doesn't do the extra validation. If all
good it does the extra business validation.