If the comment of an invoice is not complete (category and text): avoids HTTP 500 where it tried to save the comment
The UI experience is not ideal: it says that the invoices are not saved
and the user needs to click on "View Comments" to see the error.
Probably the easier way to improve this is to make the field category
mandatory when there is a text and other way round to avoid the
submission.

Using this solution at least there is no HTTP 500. I will open a ticket
for the client side improvement of the UI.